update to optimized instrumentation remove previous version
