+++ otel/opentelemetry-instrumentation-langchain/opentelemetry/instrumentation/langchain/otel_lib/tracer.py
@@ -414,13 +414,13 @@ class OpenInferenceTracer(BaseTracer):  # type: ignore
         modelmetrics = []
         try:
             self._convert_run_to_spans(run.dict(), modelmetrics)
-            token_counter = self.meter.create_up_down_counter(f"llm.request.token")
+            token_avg_counter = self.meter.create_up_down_counter(f"llm.request.token.avg")
             call_counter = self.meter.create_up_down_counter(f"llm.request.count")
-            duration_counter = self.meter.create_up_down_counter(f"llm.response.duration.avg")
+            duration_avg_counter = self.meter.create_up_down_counter(f"llm.response.duration.avg")
             for modelmetric in modelmetrics:
-                token_counter.add(modelmetric.token, {"model_id": modelmetric.model, "llm_platform": modelmetric.llm_platform})
+                token_avg_counter.add(int(modelmetric.token/modelmetric.request_count), {"model_id": modelmetric.model, "llm_platform": modelmetric.llm_platform})
                 call_counter.add(modelmetric.request_count, {"model_id": modelmetric.model, "llm_platform": modelmetric.llm_platform})
-                duration_counter.add(modelmetric.duration/modelmetric.request_count, {"model_id": modelmetric.model, "llm_platform": modelmetric.llm_platform})
+                duration_avg_counter.add(int(modelmetric.duration/modelmetric.request_count), {"model_id": modelmetric.model, "llm_platform": modelmetric.llm_platform})
         except Exception:
             logger.exception("Failed to convert run to spans")
